Fix gitconfig color settings
Prior to this commit, we had a totally messed up config for git colors.

ui is the master switch, which by default is auto. Auto means that
colors are output to the terminal but not to pipes or files. There are
verb specific color settings which we then set to auto, overriding the
ui setting.

We actually want ui=auto, which is the default, so we can just get rid
of all of this.